Don't aim for no target.
Remove additional dstReached publishing. It's redundant if the check is performed on the track node. Thus, dstReached and epsilon are best handled by the track node itself anyway. Everything above pure control moved to the track node now, and as a result we dont't have to drive to a default target while the track node is not ready.